+++ macros/src/lib.rs
@@ -247,6 +247,38 @@ pub fn part_writer(input: TokenStream) -> TokenStream {
   part_state::gen_part_wrier(input.into(), &mut DollarRefsCtx::top_level()).into()
 }
 
+/// The `split_writer` macro creates a split writer from a reference of a
+/// writer.
+///
+/// This macro specifically accepts simple expressions to indicate the partial
+/// of the writer, as shown in the following patterns:
+///
+/// - For a field: `split_writer!(&mut writer.xxx)`
+/// - For a method returning a reference: `split_writer!(writer.xxx())`.
+///
+/// Since it operates on a writer and not a state reference of the writer, the
+/// use of `$` is unnecessary.
+#[proc_macro]
+pub fn split_writer(input: TokenStream) -> TokenStream {
+  part_state::gen_split_wrier(input.into(), &mut DollarRefsCtx::top_level()).into()
+}
+
+/// The `map_watcher` macro creates a partial watcher from a reference of a
+/// watcher.
+///
+/// This macro specifically accepts simple expressions to indicate the partial
+/// of the watcher, as shown in the following patterns:
+///
+/// - For a field: `map_watcher!(&watcher.xxx)`
+/// - For a method returning a reference: `map_watcher!(watcher.xxx())`.
+///
+/// Since it operates on a watcher and not a state reference of the watcher, the
+/// use of `$` is unnecessary.
+#[proc_macro]
+pub fn map_watcher(input: TokenStream) -> TokenStream {
+  part_state::gen_part_watcher(input.into(), &mut DollarRefsCtx::top_level()).into()
+}
+
 /// The `part_reader` macro creates a partial reader from a reference of a
 /// reader.
 ///